MGS-4017 [#ccc] Fix GPU hang due to invalid dma_mask
Commit 4b5be59d31c1f45bb used the platform_driver_register() to set
dma_mask to (1 << 32) - 1 and the driver will reset it to ~0ULL. Problem is,
of_dma_configure() will check if dma_mask is set and upon
module reloading this dma_mask will point to invalid memory set
initially. Resetting it will allow to reload the module and also
set the correct dma_mask.

Signed-off-by: Marius Vlad <marius-cristian.vlad@nxp.com>