MGS-4326 [#ccc] support YUY2 to NV12/YV12/NV16
Support YUY2 to NV12/YV12/NV16 with 2D multi-dst feature.

For YUY2 to NV12 conversion:
1)	YUY2 format:  YY + UV    bpp = 16,
2)	NV12 format: YYYY + UV,   4 points = 48 bits, actual bpp = 12
there must be a loss conversion from YUY2 to NV12. The amount of data for UV(NV12) will be half as UV(YUY2).
the hardware add the value of two adjacent U(or V, YUY2) and divides by 2 as a result(U or V,NV12).

This feature can be supported on 6QP/7ULP/M845S and newer 2D cores.

Signed-off-by: Xianzhong <xianzhong.li@nxp.com>