MA-13067-2 [#imx-1215] Fix dEQP-VK.robustness* CTS fail on 8qm
CL186613 Fix some issues for
dEQP-VK.robustness.buffer_access.fragment.texel_copy.r32g32b32a32_sfloat.oob_uniform_read.range_3_texels
on v60. When linking a lib function A, if there is
another function B who is used in this lib function, we
need to update its call instruction node, but there is a
bug: If the B function is already used in the main
shader, when we update the call instruction node, we will
use the exist function parameter to find the argument
assignments, which is wrong, we should use the function
parameter from lib shader to do the search and replacement.

Change-Id: Ie60c443ee87843443d79d0fc05b77653dea20970
Signed-off-by: Richard Liu <xuegang.liu@nxp.com>