MGS-4485 [#1233] Can not pass ES32 CTS
CL186813 ReFix the bug21996, add a safe workround for this
 issue. Before unbind egl surface with wl_surface, the app
 destroies wl_surface. The destroy function can't get
 frame_sync event from server, which is bound to the
 surface, and then falls into dead loop.

For Apps like this, the user wants to find the safe workround cause a lot of Apps are open source and is not maintained by the user.

Here is the safe solution:
When destroy surface, frame_sync event is not back after 5-times check and send a display sync to server.
When display sync event is answered by the server, we can ignore frame_sync waiting if frame_sync is not back. display sync event is back to client,
On this condition, it can go on.

Signed-off-by: Yong Gan <yong.gan@nxp.com>