MGS-5088 [#ccc] QNX - signal lifecycle patch
This change is based on feedback from QNX. Not all events are
passed to kernel, so testing the Pending value to be nonzero to delete
the event may cause memory leak when event is used only on client side.
This change issues the signal pending immediately when it is being
shared with kernel. Issuing the signal pending later upon commit will
introduce delay due to commit worker and may cause memory fault in case
the event is removed by client before kernel signals it.

Date: Aug-2, 2019
Signed-off-by: Petr Cach <petr.cach@nxp.com>
Reviewed by: Jozef Bastek <jozef.bastek@nxp.com>
Reviewed by: Xianzhong Li <xianzhong.li@nxp.com>