CL277448 Fix some issues when uploading the UBO. 1) For a chip can support the unified constant, we need to use the total constant register count to check if we can upload the UBO for a shader. 2) Upload the UBO for the fragment shader first so we can get better performance.
Fix some issues when uploading the UBO: For a dynamic indexing LOAD, if any parent of this block member is an array, we can't tell if the array index is dynamically. And since so far we can't map the memory offset to the register offset(because in RA we always use packed layout to allocate), so we just don't change this LOAD.
Add a link flag to force the floating MAD instruction.
Use some built-in attributes to replace the reserved registers. Step 1: Add some built-in attributes.
Use some built-in attributes to replace the reserved registers. Step 2: Store primitiveId/perPatchAddr in TCS only when it is used in TES.
Save the per-vertex input for TES.
If there is not per-vertex input or the per-vertex input is not used in TES, we don't need to allocate the registers for these inputs.
Use some built-in attributes to replace the reserved registers. Step 3: if per-vertex input/output is not used in TCS, we don't need to allocate registers for them.
Do not initialize the per-vertex load here because the vertex count may be changed in the second linkage.