BUG27076 BUG26832 BUG27020 CL277471 Fix the crash for "./test_half -w vload_half". Do not program the shared memory if currWorkGrpNum may be changed by driver or the size is not fixed.
Add a new link type to set the fixed workGroupSize, and check whether the shader depends on the workGroupSize.
There are two situation that this CS depends on WorkGroupSize, and when application set a smaller WorkGroupSize, we need to trigger a recompilation. 1) Use WorkGroupSize to calculate the maximum register count and use register spill. With a smaller WorkGroupSize, we can use more registers and maybe we can avoid the register spill. 2) Use the local memory. With a smaller WorkGroupSize, the concurrent WorkGroupCount will be larger, we need to re-calcualte this.
Fix [escts3.2.5 / android_cts]: gles31 failed 6 cases on imx8qm_qnx / imx815 android Q.
So far we only support Vulkan and OCL to program simulated shared (local) memory.
Fix [vkcts1.1.6]oom when run "dEQP-VK.spirv_assembly.instruction.graphics.spirv_ids_abuse.sparse_ids_vert" on 815 wayland. Some refinements for reduce the size of structure "SpvSymDescriptor" and "SpvConstDescriptor".
Fix [opencv_test_core]:more than 100 fail cases when run it on 815/865 wayland. 1) When getting the symbol type ID, process the array type and the pointer type. 2) When enable/disable the OpenCV ENV config, we need to change the value of "clgDefaultPlatform->vscCoreSysCtx.hwCfg" so compiler can get the corresponding setting.